Finally got highlightjs working
Stupid vim likes to highlight html in unhelpful ways. That is: I forgot the opening quote, but my vim highlights that text the same regardless, so I didn't realize I was missing the quote before 'stylesheet', which made things break. Thanks StackOverflow (http://stackoverflow.com/questions/35388261/heroku-python-and-highlight-js-not-working/35388354#35388354) for helping me find this