Adding mostly-empty book-related page
As a side-note, I'm not a huge fan of the fact that I have to change 3 separate files when I want to simply add a link. Part of me thinks the separation is good in the long run, but on the other hand - creating a new view function requires that the url be referenced, and that the base.html point to it somehow. So it just feels fragmented (and took multiple tries to get correct)